\name{slp-package}
\alias{slp-package}
\alias{Slepians}
\docType{package}
\title{
    slp: Discrete Prolate Spheroidal (Slepian) Sequence Regression Smoothers 
}
\description{
  Interface for creation of 'slp' class smoother objects for use in 
<<<<<<< HEAD
  Generalized Additive Models as implemented by packages gam and mgcv.
=======
  Generalized Additive Models (e.g., packages 'gam' and 'mgcv'). 
>>>>>>> 377af2e41fed81b229176e9b10ff37c46e61f34b
}
\details{
\tabular{ll}{
Package: \tab slp\cr
Type: \tab Package\cr
<<<<<<< HEAD
Version: \tab 1.0-1\cr
Date: \tab 2014-07-17\cr
License: \tab GPL-2 \cr
=======
Version: \tab 1.0\cr
Date: \tab 2014-03-17\cr
License: \tab MIT \cr
>>>>>>> 377af2e41fed81b229176e9b10ff37c46e61f34b
}
}
\author{
Wesley Burr

<<<<<<< HEAD
Maintainer: Wesley Burr <wesley.burr@gmail.com>
}
\references{
  Thomson, D.J. (1982) Spectrum estimation and harmonic analysis. 
  \emph{Proceedings of the IEEE} Volume \bold{70}, number 9, pp. 1055--1096.

  Thomson, D.J. (2001) Inverse Constrained Projection Filters. 
  \emph{Proc. SPIE 4478}, Wavelets: Applications in Signal and Image Processing IX, 172 
  (December 5, 2001); doi:10.1117/12.449708
=======
with contributions from Karim Rahim and David J. Thomson (in particular the
compiled Fortran code for dpss generation).

Maintainer: Wesley Burr <wesley.burr@gmail.com>
}
\references{
  Thomson, D.J (1982)
  Spectrum estimation and harmonic analysis. \emph{Proceedings of the IEEE}
  Volume \bold{70}, number 9, pp. 1055--1096.
>>>>>>> 377af2e41fed81b229176e9b10ff37c46e61f34b

  Hastie, T. J. (1991)
  \emph{Generalized additive models.}
  Chapter 7 of \emph{Statistical Models in S}
  eds J. M. Chambers and T. J. Hastie, Wadsworth \& Brooks/Cole.

  Hastie, T. and Tibshirani, R. (1990)
  \emph{Generalized Additive Models.}
  London: Chapman and Hall.

  Wood, S.N. (2011) Fast stable restricted maximum likelihood 
  and marginal likelihood estimation of semiparametric generalized linear 
  models. Journal of the Royal Statistical Society (B) 73(1):3-36

  Wood, S.N. (2004) Stable and efficient multiple smoothing parameter estimation for
  generalized additive models. J. Amer. Statist. Ass. 99:673-686. 

  Wood, S.N. (2006) \emph{Generalized Additive Models: an introduction with R},
  CRC
}
<<<<<<< HEAD
=======
\keyword{algebra}
>>>>>>> 377af2e41fed81b229176e9b10ff37c46e61f34b
\keyword{models}
\keyword{regression}
\keyword{ts}
\keyword{smooth}
<<<<<<< HEAD
\keyword{Slepian}
\seealso{
\code{\link[multitaper:spec.mtm]{multitaper}}
\code{\link[mgcv]{mgcv}}
\code{\link[gam]{gam}}
=======
\seealso{
\code{\link[multitaper:spec.mtm]{multitaper}}
\code{\link[mgcv:mgcv-package]{mgcv}}
\code{\link[gam:gam]{gam}}
>>>>>>> 377af2e41fed81b229176e9b10ff37c46e61f34b
}

