\name{slp-package}
\alias{slp-package}
\alias{slp.package}
\alias{Slepians}
\docType{package}
\title{
    slp: Discrete Prolate Spheroidal (Slepian) Sequence Regression Smoothers 
}
\description{
  Interface for creation of 'slp' class smoother objects for use in 
  Generalized Additive Models as implemented by packages gam and mgcv.
}
\details{
\tabular{ll}{
Package: \tab slp\cr
Type: \tab Package\cr
Version: \tab 1.0-5\cr
Date: \tab 2016-08-28\cr
License: \tab GPL-2 \cr
}
}
\author{
Wesley Burr, with contributions from Karim Rahim

Maintainer: Wesley Burr <wesley.burr@gmail.com>
}
\references{
  Thomson, D.J. (1982) Spectrum estimation and harmonic analysis. 
  \emph{Proceedings of the IEEE} Volume \bold{70}, number 9, pp. 1055--1096.

  Thomson, D.J. (2001) Inverse Constrained Projection Filters. 
  \emph{Proc. SPIE 4478}, Wavelets: Applications in Signal and Image Processing IX, 172 
  (December 5, 2001); doi:10.1117/12.449708

  Hastie, T. J. (1991)
  \emph{Generalized additive models.}
  Chapter 7 of \emph{Statistical Models in S}
  eds J. M. Chambers and T. J. Hastie, Wadsworth \& Brooks/Cole.

  Hastie, T. and Tibshirani, R. (1990)
  \emph{Generalized Additive Models.}
  London: Chapman and Hall.

  Wood, S.N. (2011) Fast stable restricted maximum likelihood 
  and marginal likelihood estimation of semiparametric generalized linear 
  models. Journal of the Royal Statistical Society (B) 73(1):3-36

  Wood, S.N. (2004) Stable and efficient multiple smoothing parameter estimation for
  generalized additive models. J. Amer. Statist. Ass. 99:673-686. 

  Wood, S.N. (2006) \emph{Generalized Additive Models: an introduction with R},
  CRC
}
\keyword{models}
\keyword{regression}
\keyword{ts}
\keyword{smooth}
\keyword{Slepian}

